Add files via upload
This is a first test project on European vanilla option pricing using Monte Carlo. 
This only includes the call and put options, although the formulas can be generalized to any payoff function g. 
Also analytical formulas for the call and put are provided to make sure we obtain correct results. 
To upload the file, do a pull request.